seperate decleration and implementation for logger
