+++ CMakeLists.txt
@@ -25,7 +25,8 @@ find_package(OpenCV REQUIRED)
 find_package(fmt REQUIRED)
 
 add_library(tensorrt_cpp_api SHARED
-        src/engine.cpp)
+        src/engine.cpp
+        src/logger.cpp)
 
 target_include_directories(tensorrt_cpp_api PUBLIC ${OpenCV_INCLUDE_DIRS} ${CUDA_INCLUDE_DIRS} ${TensorRT_INCLUDE_DIRS} include)
 target_link_libraries(tensorrt_cpp_api PUBLIC ${OpenCV_LIBS} ${CUDA_LIBRARIES} ${CMAKE_THREAD_LIBS_INIT} ${TensorRT_LIBRARIES} fmt::fmt)
